smart-pharmacy-agent/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
pydantic==2.9.2
SQLAlchemy==2.0.36
pandas==2.2.2
numpy==1.26.4
haversine==2.8.1
streamlit==1.38.0
pytest==8.3.2
groq==0.11.0
python-dotenv==1.0.1
httpx==0.27.0
